feature:【engine】字节码执行引擎添加引用类型追踪

+++ vm/jmethod_area.h
@@ -18,7 +18,7 @@ class jmethod_area
 private:
     char *m_buff;
     int m_capacity;
-    std::atomic<int> m_offset;
+    std::atomic<size_t> m_offset;
 
 public:
     static constexpr int DEFAULT_CAPACITY = 1 * 1024 * 1024;
@@ -32,7 +32,7 @@ public:
      * @param len 需要的大小，单位是字节。为 0 时返回 nullptr
      * @return 成功时返回指针，失败退出虚拟机。保证返回的指针对齐到 8 字节
      */
-    void *malloc_bytes(int len);
+    void *malloc_bytes(size_t len);
 
 
     template<typename T>
@@ -43,7 +43,7 @@ public:
     }
 
     template<typename T>
-    T* calloc_type(int n)
+    T* calloc_type(size_t n)
     {
         if (n == 0) return nullptr;
